refactor: eliminar HasRoles y HasApiTokens de servicios que no lo necesitaban y crear servicio para controladores de liquidación

+++ app/Services/CompanyService.php
@@ -135,6 +135,10 @@ class CompanyService
             ]);
 
             if ($request->hasFile('logo')) {
+                $validationResponse = $this->validateLogo($request);
+                if ($validationResponse !== true) {
+                    return $validationResponse;
+                }
                 if ($company->logo_path) {
                     Helper::deleteFile($company->logo_path);
                 }
